Update archlinux-keyring before pacstrapping
There have occasionally been hard-to-diagnose issues with packages
failing checksum tests during the pacstrap installation process. The
check for pacman-init directly above this code was added in response to
these failures and seemed to address the problem.

The issue reappeared and happened even after waiting for pacman-init to
reach active state. This update of archlinux-keyring seems to solve
the problem for now.